# Fix performance regression in 3 or more qubits RB