@itoko 2023-01-19
Close as this fix is included in #982 (via #922)